feat: Code refactoring for better readability
