+++ app/templates/index.html
@@ -9,7 +9,7 @@
         <h2>Database Connection</h2>
         <button
             class="btn btn-success"
-            hx-get="/web/test-connection"
+            hx-get="/test-connection"
             hx-target="#connection-status"
             hx-indicator="#test-spinner">
             Test Connection
@@ -39,14 +39,14 @@
         <h2>Users</h2>
         <button
             class="btn btn-primary"
-            hx-get="/web/users"
+            hx-get="/users"
             hx-target="#users-container"
             hx-indicator="#load-spinner">
             Refresh
             <span id="load-spinner" class="htmx-indicator spinner"></span>
         </button>
     </div>
-    <div id="users-container" hx-get="/web/users" hx-trigger="load">
+    <div id="users-container" hx-get="/users" hx-trigger="load">
         <div class="empty-state">
             <div class="empty-state-icon">⏳</div>
             <div class="empty-state-text">Loading users...</div>
